Rename serviceUtil to Utils
Change name to utils.js
